add urdf, rviz config
