+++ sick_wagen/config/whill/whill_params.yaml
@@ -1,0 +1,6 @@
+whill_modelc_controller_node:
+  ros__parameters:
+    serialport: "/dev/ttyUSB-WHILL"
+    wheel_radius: 0.135
+    send_interval: 10
+    enable_cmd_vel_control: true
